+++ tests/hikari/test_channels.py
@@ -112,6 +112,9 @@ class TestPartialChannel:
         model.name = None
         assert str(model) == "Unnamed PartialChannel ID 1234567"
 
+    def test_mention_property(self, model):
+        assert model.mention == "<#1234567>"
+
     @pytest.mark.asyncio()
     async def test_delete(self, model):
         model.app.rest.delete_channel = mock.AsyncMock()
@@ -335,12 +338,9 @@ class TestGuildChannel:
         model.app.shard_count = 3
         assert model.shard_id == 2
 
-    def test_mention_property(self, model):
-        assert model.mention == "<#69420>"
-
     @pytest.mark.asyncio()
     async def test_edit_overwrite(self, model):
-        model.app.rest.edit_permission_overwrites = mock.AsyncMock()
+        model.app.rest.edit_permission_overwrite = mock.AsyncMock()
         user = mock.Mock(users.PartialUser)
         await model.edit_overwrite(
             333,
@@ -350,7 +350,7 @@ class TestGuildChannel:
             reason="vrooom vroom",
         )
 
-        model.app.rest.edit_permission_overwrites.assert_called_once_with(
+        model.app.rest.edit_permission_overwrite.assert_called_once_with(
             69420,
             333,
             target_type=user,
@@ -361,13 +361,13 @@ class TestGuildChannel:
 
     @pytest.mark.asyncio()
     async def test_edit_overwrite_target_type_none(self, model):
-        model.app.rest.edit_permission_overwrites = mock.AsyncMock()
+        model.app.rest.edit_permission_overwrite = mock.AsyncMock()
         user = mock.Mock(users.PartialUser)
         await model.edit_overwrite(
             user, allow=permissions.Permissions.BAN_MEMBERS, deny=permissions.Permissions.CONNECT, reason="vrooom vroom"
         )
 
-        model.app.rest.edit_permission_overwrites.assert_called_once_with(
+        model.app.rest.edit_permission_overwrite.assert_called_once_with(
             69420,
             user,
             allow=permissions.Permissions.BAN_MEMBERS,